requirements.txt
Flask==2.0.2
gunicorn
qrcode==7.3.1
opencv-python==4.5.5.62
Pillow==8.4.0